Refactor project structure and update build configs
Moved core source files to src/pdf-parse and node-specific files to src/node. Updated build outputs, TypeScript configs, and package.json exports to reflect new directory structure. Renamed utility configs and extractor configs for clarity. Updated test imports and removed legacy TableUtil and related test. Adjusted Vite and Vitest configs for new paths. Expanded API documentation for geometry and table types.
