+++ .sonarcloud.properties
@@ -20,7 +20,7 @@ sonar.sourceEncoding=UTF-8
 #sonar.sourceEncoding=UTF-8
 
 # Exclusions for copy-paste detection
-sonar.cpd.exclusions=**/dist/**,**/node_modules/**
+sonar.cpd.exclusions=**/dist/**,**/node_modules/**,**/bin/**
 
 # Coverage report - Vitest with V8 reporter writes lcov to coverage/lcov.info when run via `npm run coverage`
 # Ensure your CI generates the coverage report before running sonar-scanner
