+++ api-extractor.json
@@ -45,7 +45,7 @@
 	 *
 	 * SUPPORTED TOKENS: <projectFolder>, <packageName>, <unscopedPackageName>
 	 */
-	"mainEntryPointFilePath": "dist/esm/index.d.ts",
+	"mainEntryPointFilePath": "dist/pdf-parse/esm/index.d.ts",
 
 	/**
 	 * A list of NPM package names whose exports should be treated as part of this package.
@@ -278,7 +278,7 @@
 		 * SUPPORTED TOKENS: <projectFolder>, <packageName>, <unscopedPackageName>
 		 * DEFAULT VALUE: "<projectFolder>/dist/<unscopedPackageName>.d.ts"
 		 */
-		"untrimmedFilePath": "dist/cjs/index.d.cts"
+		"untrimmedFilePath": "dist/pdf-parse/cjs/index.d.cts"
 
 		/**
 		 * Specifies the output path for a .d.ts rollup file to be generated with trimming for an "alpha" release.
