+++ src/pdf-parse/Exception.ts
