+++ src/pdf-parse/ImageResult.ts
