+++ src/pdf-parse/InfoResult.ts
