+++ src/pdf-parse/PathGeometry.ts
