+++ src/pdf-parse/ScreenshotResult.ts
