+++ src/pdf-parse/TextResult.ts
