+++ src/pdf-parse/_TableUtil.ts
