+++ src/pdf-parse/geometry/Line.ts
