+++ src/pdf-parse/geometry/LineStore.ts
