+++ src/pdf-parse/geometry/Point.ts
