+++ src/pdf-parse/geometry/Rectangle.ts
