+++ src/pdf-parse/geometry/Shape.ts
