+++ src/pdf-parse/index.browser.ts
