+++ tests/unit/test-geometry/geometry-table.full.test.ts
@@ -1,8 +1,5 @@
-﻿import { describe, expect, test } from 'vitest';
-
-import { Line } from '../../../src/geometry/Line';
-import { Point } from '../../../src/geometry/Point';
-import { Table } from '../../../src/geometry/Table';
+﻿import { Line, Point, Table } from 'pdf-parse';
+import { describe, expect, test } from 'vitest';
 
 describe('Table constructor tests', (): void => {
 	test('table constructor with horizontal line', (): void => {
