--- tsconfig.utils.json
@@ -1,25 +1,0 @@
-{
-	"compilerOptions": {
-		"baseUrl": ".",
-		"rootDir": "utils",
-		"outDir": "dist/utils/esm",
-		"declaration": true,
-		"declarationMap": true,
-		"sourceMap": true,
-		"esModuleInterop": true,
-		"isolatedModules": true,
-		"target": "ES2022",
-		"module": "node16",
-		"moduleResolution": "node16",
-		"jsx": "react",
-		"noEmit": false,
-		"noUncheckedIndexedAccess": false,
-		"skipLibCheck": true,
-		"strict": true,
-		"verbatimModuleSyntax": true,
-		"paths": {
-			"pdf-parse/utils": ["utils"]
-		}
-	},
-	"include": ["utils/**/*"]
-}
