+++ vite.config.cjs.ts
@@ -2,12 +2,12 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
 	build: {
-		outDir: 'dist/cjs',
+		outDir: 'dist/pdf-parse/cjs',
 		emptyOutDir: false,
 		sourcemap: true,
 		minify: 'terser',
 		lib: {
-			entry: 'src/index.ts',
+			entry: 'src/pdf-parse/index.ts',
 			name: 'PdfParse',
 			fileName: 'index',
 			formats: ['cjs'],
