+++ vite.config.ts
@@ -4,12 +4,12 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
 	build: {
-		outDir: 'dist/browser',
+		outDir: 'dist/pdf-parse/browser',
 		emptyOutDir: true,
 		sourcemap: true,
 		minify: 'terser',
 		lib: {
-			entry: 'src/index.browser.ts',
+			entry: 'src/pdf-parse/index.browser.ts',
 			name: 'PdfParse',
 			fileName: (format) => `pdf-parse.${format}.js`,
 			formats: ['es', 'umd'],
@@ -23,9 +23,9 @@ export default defineConfig({
 				const source = join(cwd, 'node_modules/pdfjs-dist/legacy/build', 'pdf.worker.mjs');
 				const source_min = join(cwd, 'node_modules/pdfjs-dist/legacy/build', 'pdf.worker.min.mjs');
 				const source_map = join(cwd, 'node_modules/pdfjs-dist/legacy/build', 'pdf.worker.mjs.map');
-				const dest_browser = join(cwd, 'dist/browser/');
-				const dest_esm = join(cwd, 'dist/esm/');
-				const dest_cjs = join(cwd, 'dist/cjs/');
+				const dest_browser = join(cwd, 'dist/pdf-parse/browser/');
+				const dest_esm = join(cwd, 'dist/pdf-parse/esm/');
+				const dest_cjs = join(cwd, 'dist/pdf-parse/cjs/');
 
 				cpSync(source, join(dest_browser, basename(source)));
 				cpSync(source_map, join(dest_browser, basename(source_map)));
@@ -38,7 +38,7 @@ export default defineConfig({
 		{
 			name: 'copy-dist-to-reports',
 			closeBundle() {
-				const source = join(process.cwd(), 'dist/browser');
+				const source = join(process.cwd(), 'dist/pdf-parse/browser');
 				const dest = join(process.cwd(), 'reports/demo/dist-browser');
 				cpSync(source, dest, { recursive: true });
 			},
